Create deployment.py
